requirements.txt
flask==2.0.1
werkzeug==2.0.3
python-dotenv==0.19.0
flask-cors==3.0.10
pymongo==3.12.0
gunicorn==20.1.0
python-decouple==3.4
requests==2.31.0 